refactor(proxy): extract background acknowledgement check
